moved block statement parser function in statements folder
