+++ internal/ast/statements/block_statement.go
@@ -1,4 +1,4 @@
-package expressions
+package statements
 
 import (
 	"bytes"
